requirements.txt
pdfplumber==0.10.3
pandas==2.2.0
openpyxl==3.1.2
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0
Flask==3.0.0
Werkzeug==3.0.1 